requirements.txt
requests==2.31.0
aiohttp==3.9.1
loguru==0.7.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
google-cloud-artifact-registry==1.10.0